requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23

# OpenAI Integration
openai==1.6.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1

# Email
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
schedule==1.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
